==> notebooks/1_preprocessing.ipynb <==
import pandas as pd
import numpy as np

## Качество питьевой воды

wp_df = pd.read_csv('water_potability.csv')

wp_df

wp_df.describe()

wp_df.info()

wp_df.isnull().sum()

wp_df.fillna(wp_df.mean(), inplace=True)

wp_df

wp_df.to_csv('water_potability_preprocessed.csv')

## Автомобили

names = ['symboling', 'normalized-losses', 'make', 'fuel-type', 'aspiration', 'num-of-doors', 'body-style', 'drive-wheels',
         'engine-location', 'wheel-base', 'length', 'width', 'height', 'curb-weight', 'engine-type', 'num-of-cylinders',
         'engine-size', 'fuel-system', 'bore', 'stroke', 'compression-ratio', 'horsepower', 'peak-rpm', 'city-mpg',
         'highway-mpg', 'price']

auto_df = pd.read_csv('automobile.csv', names=names)

auto_df

auto_df.describe()

Выведем статистику для всех признаков

auto_df.describe(include='all')

auto_df.info()

auto_df = auto_df.replace('?', np.nan)

auto_df.isnull().sum()

auto_df.dtypes

auto_df['normalized-losses'] = auto_df['normalized-losses'].astype('float64')
auto_df['normalized-losses'].fillna(auto_df['normalized-losses'].mean(), inplace=True)

auto_df.dtypes

auto_df.dropna(inplace=True)

auto_df

auto_df[['bore', 'stroke']] = auto_df[['bore', 'stroke']].astype('float64')
auto_df[['normalized-losses', 'horsepower', 'peak-rpm', 'price']] = auto_df[
    ['normalized-losses', 'horsepower', 'peak-rpm', 'price']].astype('int64')

auto_df.info()

pd.unique(auto_df['body-style'])

pd.unique(auto_df['make'])

from sklearn.preprocessing import OrdinalEncoder

Закодируем признак 'make' (марки авто) с помощью OrdinalEncoder()

enc = OrdinalEncoder()
auto_df['make'] = enc.fit_transform(auto_df[["make"]])

Кодируем категориальные признаки

auto_df = pd.get_dummies(auto_df, columns=
               ['body-style', 'drive-wheels', 'engine-type', 'num-of-cylinders', 'fuel-system'],
               prefix=['bs', 'dw', 'et', 'num', 'fs'])

Закодируем некоторые признаки при помощи словаря и применим метод map

map_fuel_type = {'gas': 0, 'diesel': 1}
map_aspiration = {'std': 0, 'turbo': 1}
map_num_of_doors = {'four': 0, 'two': 1}
map_engine_loc = {'front': 0, 'rear': 1}

auto_df['fuel-type'] = auto_df['fuel-type'].map(map_fuel_type)
auto_df['aspiration'] = auto_df['aspiration'].map(map_aspiration)
auto_df['num-of-doors'] = auto_df['num-of-doors'].map(map_num_of_doors)
auto_df['engine-location'] = auto_df['engine-location'].map(map_engine_loc)

auto_df

auto_df.to_csv('automobile_preprocessed.csv')
