==> notebooks/2_classification.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler

data = pd.read_csv("../data/water_potability_preprocessed.csv")

data

data = data.iloc[:, 1:]

X = data.iloc[:, :-1].values # все признаки, кроме целевого
y = data['Potability'].values # целевой признак

разделяем выборки на обучающие и тестовые

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

X_train

**Метод k-ближайших соседей (K-Nearest Neighbors)**

from sklearn.neighbors import KNeighborsClassifier

knn = KNeighborsClassifier()
params = {'n_neighbors': np.arange(1,6,1)}
clf = GridSearchCV(knn, params)
clf.fit(X_train, y_train)
clf.best_params_

from sklearn.metrics import classification_report, confusion_matrix

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

своя реализация kNN

def euclidian_metric(a, b):  # реализуем Евклидову метрику через норму
    return np.sqrt(np.sum((a - b) ** 2, axis=1))

def find_neighbours(k, y, distances):
    neighbours = y[np.argsort(distances)][:k]
    neighbours_distances = distances[neighbours][:k]

    return neighbours, neighbours_distances

def get_closest_classes(neighbours):
    arr = np.unique(neighbours, return_counts=True)
    best_classes = arr[0][arr[1] == np.amax(arr[1])]

    return best_classes

def choose_best_class(best_classes, neighbours, neighbouring_distances):
    min_mean_dist = np.inf
    best_class = None

    for elem in np.unique(neighbours):
        mean_dist = np.mean(neighbouring_distances[neighbours == elem])
        if mean_dist < min_mean_dist:
            best_class = elem
            min_mean_dist = mean_dist

    return best_class

def nearest_neighbours_classify(x, y, k, x_pred):
    res = np.zeros(x_pred.shape[0], dtype=y.dtype)

    for i in range(x_pred.shape[0]):
        distances = euclidian_metric(x, x_pred[i])  # считаем расстояния до классов

        neighbours, neighbouring_distances = find_neighbours(k, y, distances)  # находим ровно k соседей этой точки
            
        best_classes = get_closest_classes(neighbours)  # обнаруживаем классы, которые имеются среди соседей

        res[i] = choose_best_class(best_classes, neighbours, neighbouring_distances)  # выбираем наиболее релевантный класс по среднему расстоянию до него среди соседей
    return res

y_pred = nearest_neighbours_classify(X_train, y_train, clf.best_params_['n_neighbors'], X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

результат получился одинаковым

**Классификатор дерева решений (Decision Tree Classifier)**

from sklearn import tree

dtc = tree.DecisionTreeClassifier()
params = {'max_depth': np.arange(1,15,1),
          'max_features': np.arange(5,X.shape[1]-1,1)}
clf = GridSearchCV(dtc, params)
clf = clf.fit(X_train, y_train)
clf.best_params_

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

**Наивный байесовский классификатор (Naive Bayes)**

from sklearn.naive_bayes import GaussianNB, BernoulliNB

g_nb = GaussianNB()
b_nb = BernoulliNB()

g_nb.fit(X_train, y_train)

y_pred = g_nb.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(g_nb.score(X_test, y_test))

b_nb.fit(X_train, y_train)

y_pred = b_nb.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(b_nb.score(X_test, y_test))

Наивный байесовский классификатор с распределением Бернулли показал лучший результат по сравнению с нормальным распределением

**Метод опорных векторов (Support Vector Machines)**

from sklearn import svm

svc = svm.SVC()
params = {'C': np.arange(0.1,1.1,0.1),
          'degree': np.arange(2,5,1)}
clf = GridSearchCV(svc, params)
clf.fit(X_train, y_train)
clf.best_params_

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

**Логистическая регрессия (Logistic Regression)**

from sklearn.linear_model import LogisticRegression

logReg = LogisticRegression()
params = {'C': np.arange(0.1,1.1,0.1)}
clf = GridSearchCV(logReg, params)
clf.fit(X_train, y_train)
clf.best_params_

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

**Perceptron**

from sklearn.linear_model import Perceptron

perc = Perceptron()
params = {'alpha': np.array([0.0001, 0.0005, 0.001, 0.005, 0.01, 0.05, 0.1, 0.5]),
          'l1_ratio': np.arange(0,1.05,0.05)}
clf = GridSearchCV(perc, params)
clf.fit(X_train, y_train)
clf.best_params_

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))
print(clf.score(X_test, y_test))

**Вывод:** самый лучший классификатор по точности - SVM
