==> notebooks/3_regression.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler, OrdinalEncoder

from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error

import warnings
warnings.filterwarnings("ignore")

data = pd.read_csv('../data/automobile_preprocessed.csv')

data = data.iloc[:, 1:]

enc = OrdinalEncoder()
data['make'] = enc.fit_transform(data[["make"]])

data

data.info()

X = data.drop(columns=['price']).values # все признаки, кроме целевого
y = data['price'].values # целевой признак

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=42)

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

X_train

def print_metrics(y_test, y_pred):
    print("R2:", r2_score(y_test, y_pred))
    print("MSE:", mean_squared_error(y_test, y_pred))
    print("RMSE:", mean_squared_error(y_test, y_pred, squared=False))
    print("MAE:", mean_absolute_error(y_test, y_pred))

**Linreg**

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LinearRegression.html#sklearn.linear_model.LinearRegression

from sklearn.linear_model import LinearRegression

lr = LinearRegression()
lr.fit(X_train, y_train)

print_metrics(y_test, lr.predict(X_test))

своя реализация LinReg

class linReg:
    def __init__(self, lr = 0.0001, iters = 5000):
        self.lr = lr
        self.iters = iters

    def fit(self, x, y):
        x = np.c_[np.ones(x.shape[0]), x]
        self.w = np.random.rand((x.shape[1]))

        for i in range(self.iters):
            self.w = self.w - self.lr * np.dot(x.T, (np.dot(x, self.w) - y)) * 2 / x.shape[0]

    def predict(self, x):
        x = np.c_[np.ones(x.shape[0]), x]
        return np.dot(x, self.w)

lr_ = linReg()
lr_.fit(X_train, y_train)
print_metrics(y_test, lr_.predict(X_test))

Увеличим количество итераций

lr_ = linReg(iters=50000)
lr_.fit(X_train, y_train)

print_metrics(y_test, lr_.predict(X_test))

Качество готовой модели и собственной могут отличаться. В собственной модели от величины шага и кол-ва итераций зависят правильность построения модели

**DecisionTreeRegressor**

https://scikit-learn.org/stable/modules/generated/sklearn.tree.DecisionTreeRegressor.html#sklearn.tree.DecisionTreeRegressor

from sklearn.tree import DecisionTreeRegressor

dtr = DecisionTreeRegressor(random_state=42)
params = {'max_depth': np.arange(1,15,1),
          'max_features': np.arange(5,X.shape[1]-1,1)}
clf = GridSearchCV(dtr, params)
clf.fit(X_train, y_train)
clf.best_params_

print_metrics(y_test, clf.predict(X_test))

Качество данной модели немного хуже, чем в предыдущей

**Lasso**

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Lasso.html#sklearn.linear_model.Lasso

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LassoCV.html#sklearn.linear_model.LassoCV

from sklearn.linear_model import Lasso, LassoCV

lasso = Lasso(random_state=42)
params = {'alpha': np.arange(0.1,2.05,0.05)}
clf = GridSearchCV(lasso, params)
clf.fit(X_train, y_train)
clf.best_params_

print_metrics(y_test, clf.predict(X_test))

Lasso с кросс-валидацией

lasso_cv = LassoCV(random_state=42)
lasso_cv.fit(X_train, y_train)

print_metrics(y_test, lasso_cv.predict(X_test))

Lasso и LassoCV показывают примерно одинаковое качество.

**Ridge**

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html#sklearn.linear_model.Ridge

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.RidgeCV.html#sklearn.linear_model.RidgeCV

from sklearn.linear_model import Ridge, RidgeCV

ridge = Ridge(random_state=42)
params = {'alpha': np.arange(0.1,2.05,0.05)}
clf = GridSearchCV(ridge, params)
clf.fit(X_train, y_train)
clf.best_params_

print_metrics(y_test, clf.predict(X_test))

Ridge с кросс-валидацией

ridge_cv = RidgeCV(cv=5)
ridge_cv.fit(X_train, y_train)

print_metrics(y_test, ridge_cv.predict(X_test))

Тоже самое и с Ridge и RidgeCV

**Elastic Net**

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNet.html#sklearn.linear_model.ElasticNet

https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.ElasticNetCV.html#sklearn.linear_model.ElasticNetCV

from sklearn.linear_model import ElasticNet, ElasticNetCV

params = {'alpha': np.arange(0.1,2.05,0.05),
          'l1_ratio': np.arange(0.1,1.1,0.1)}

clf = GridSearchCV(ElasticNet(random_state=42), params)
clf.fit(X_train, y_train)
clf.best_params_

print_metrics(y_test, clf.predict(X_test))

ElasticNet без гиперпараметров

print_metrics(y_test, ElasticNet(random_state=42).fit(X_train, y_train).predict(X_test))

ElasticNet с кросс-валидацией

en_cv = ElasticNetCV(cv=5, random_state=42)
en_cv.fit(X_train, y_train)

print_metrics(y_test, en_cv.predict(X_test))
