==> notebooks/6_perceptron.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

# Task 1

class MLP:
    def __init__(self, activation='relu', hidden_layer_shape=(4,), learning_rate=0.1, max_iter=10000):
        self.X = None
        self.y = None
        self.parameters = None
        
        self.n_h = hidden_layer_shape
        self.activation = activation #sigmoid, relu
        self.learning_rate = learning_rate
        self.max_iter = max_iter
    
    def __sigmoid(self, Z):
        A = 1 / (1 + np.exp(-Z))
        cache = Z

        return A, cache

    def __relu(self, Z):
        A = np.maximum(0, Z)
        cache = Z 
        
        return A, cache
    
    def __sigmoid_backward(self, dA, cache):
        Z = cache
        A = 1 / (1 + np.exp(-Z))
        dZ = dA * A * (1 - A)

        return dZ

    def __relu_backward(self, dA, cache):
        Z = cache
        dZ = np.array(dA, copy=True)
        dZ[Z <= 0] = 0

        return dZ
    
    def __initialize_parameters(self, layer_dims):
        parameters = {}
        L = len(layer_dims)

        for l in range(1, L):
            parameters["W" + str(l)] = np.random.randn(layer_dims[l], layer_dims[l-1]) * 0.01
            parameters["b" + str(l)] = np.zeros((layer_dims[l], 1))

        return parameters
    
    def __forward_propagation(self, X, parameters):
        
        def linear_forward(A, W, b):
            Z = np.dot(W, A) + b
            cache = (A, W, b)

            return Z, cache
    
        def linear_activation_forward(A_prev, W, b, activation):
            if activation == "sigmoid":
                Z, linear_cache = linear_forward(A_prev, W, b)
                A, activation_cache = self.__sigmoid(Z)

            if activation == "relu":
                Z, linear_cache = linear_forward(A_prev, W, b)
                A, activation_cache = self.__relu(Z)

            cache = (linear_cache, activation_cache)

            return A, cache
    
        caches = []
        A = X.T
        L = len(parameters) // 2

        for l in range(1, L):
            A_prev = A 

            A, cache = linear_activation_forward(A_prev, parameters['W' + str(l)], parameters['b' + str(l)], 'relu')
            caches.append(cache)

        AL, cache = linear_activation_forward(A, parameters['W' + str(L)], parameters['b' + str(L)], 'sigmoid')
        caches.append(cache)

        return AL, caches
    
    def __compute_cost(self, AL, Y):
        m = Y.shape[0]
        
        cost = -np.sum(Y * np.log(AL) + (1 - Y) * np.log(1 - AL)) / m
        cost = np.squeeze(cost)
        
        return cost
    
    def __backward_propagation(self, AL, Y, caches):

        def linear_backward(dZ, cache):
            A_prev, W, b = cache
            m = A_prev.shape[0]

            dW = np.dot(dZ, A_prev.T) / m
            db = np.sum(dZ, axis=1, keepdims=True) / m
            dA_prev = np.dot(W.T, dZ)

            return dA_prev, dW, db
    
        def linear_activation_backward(dA, cache, activation):
            linear_cache, activation_cache = cache

            if activation == "relu":
                dZ = self.__relu_backward(dA, activation_cache)
                dA_prev, dW, db = linear_backward(dZ, linear_cache)

            elif activation == "sigmoid":
                dZ = self.__sigmoid_backward(dA, activation_cache)
                dA_prev, dW, db = linear_backward(dZ, linear_cache)

            return dA_prev, dW, db
        
        grads = {}
        L = len(caches)
        m = AL.shape[0]
        Y = Y.reshape(AL.shape)        

        dAL = -(np.divide(Y, AL) - np.divide(1 - Y, 1 - AL))

        current_cache = caches[L-1]
        dA_prev_temp, dW_temp, db_temp = linear_activation_backward(dAL, current_cache, 'sigmoid')
        grads["dA" + str(L-1)] = dA_prev_temp
        grads["dW" + str(L)] = dW_temp
        grads["db" + str(L)] = db_temp

        for l in reversed(range(L-1)):
            current_cache = caches[l]
            dA_prev_temp, dW_temp, db_temp = linear_activation_backward(grads["dA" + str(l + 1)], current_cache, 'relu')
            grads["dA" + str(l)] = dA_prev_temp
            grads["dW" + str(l + 1)] = dW_temp
            grads["db" + str(l + 1)] = db_temp

        return grads
    
    def __update_parameters(self, params, grads):
        parameters = params.copy()
        L = len(parameters) // 2 # number of layers in the neural network

        for l in range(L):
            parameters["W" + str(l+1)] = parameters['W' + str(l+1)] - self.learning_rate * grads['dW' + str(l+1)]
            parameters["b" + str(l+1)] = parameters['b' + str(l+1)] - self.learning_rate * grads['db' + str(l+1)]

        return parameters
    
    def fit(self, X, y, verbose=False):
        self.X = X
        self.y = y
        
        self.n_x = X.shape[1]
        self.n_y = y.shape[1]
        
        parameters = self.__initialize_parameters([self.n_x, *self.n_h, self.n_y])
        
        for i in range(self.max_iter):
            AL, caches = self.__forward_propagation(X, parameters)
            cost = self.__compute_cost(AL, y)
            grads = self.__backward_propagation(AL, y, caches)
            parameters = self.__update_parameters(parameters, grads)
            
            if verbose: print ("Cost after iteration %i: %f" %(i, cost))
            
        self.parameters = parameters
    
    def predict(self, X):
        AL, cache = self.__forward_propagation(X, self.parameters)
        predictions = (AL > 0.5)

        return predictions

    def get_parameters(self):
        return self.parameters

titanic_df_train = pd.read_csv("../data/train.csv")
titanic_df_test = pd.read_csv("../data/test.csv")

iris_names = ['sepal length in cm', 'sepal width in cm', 'petal length in cm', 'petal width in cm', 'class']
iris_df = pd.read_csv("../data/iris.csv", names=iris_names)

titanic_df_train.info()

titanic_df_train.isnull().sum()

titanic_df = titanic_df_train.fillna(titanic_df_train.mean()).dropna()

titanic_df = titanic_df.drop(columns=["Name", "Ticket", "Cabin"])

titanic_df['Sex'] = pd.Categorical(titanic_df.Sex).codes
titanic_df['Embarked'] = pd.Categorical(titanic_df.Embarked).codes

titanic_df

X = titanic_df.drop(columns=["Survived"]).values
y = titanic_df["Survived"].values.reshape(-1, 1)

X.shape

y.reshape(y.shape[0], -1).shape

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model = MLP(max_iter=150, learning_rate=0.1)
model.fit(X_train, y_train)

model.get_parameters()

model.predict(X_test)

iris_df["class"] = pd.Categorical(iris_df["class"]).codes

iris_df

X = iris_df.drop(columns=["class"]).values
y = iris_df["class"].values.reshape(-1, 1)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

scaler = StandardScaler()

X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

model = MLP()
model.fit(X_train, y_train)

model.predict(X_test)

print(y_test)

# Task 2

water_df = pd.read_csv("../data/water_potability_preprocessed.csv").iloc[:, 1:]
auto_df = pd.read_csv('../data/automobile_preprocessed.csv').iloc[:, 1:]
